requirements.txt
autopep8==2.0.2
imageio==2.26.1
numpy==1.24.2
Pillow==9.4.0
pycodestyle==2.10.0
pygame==2.3.0
pygame-animatedgif==0.1.0
tomli==2.0.1
